almost done

+++ app/api/dashboard/route.js
@@ -4,35 +4,37 @@ import { auth } from "@/auth";
 
 export const POST = auth(async function POST(request) {
   try {
-    const req = await request.json();
+    if (request.auth.user && request.auth.user.user_type == "vendor") {
+      const req = await request.json();
 
-    // console.log("dashboard route: ",parseInt( request.auth.user.id));
-    // const vendor = await prisma.eventVendor.findUnique({
-    //   where: { id: parseInt( request.auth.user.id) }
-    // });
-    // if (!vendor) {
-    //   console.log("EventVendor not found!");
-    // } else {
-    //   console.log("Found EventVendor:", vendor);
-    // }
+      // console.log("dashboard route: ",parseInt( request.auth.user.id));
+      // const vendor = await prisma.eventVendor.findUnique({
+      //   where: { id: parseInt( request.auth.user.id) }
+      // });
+      // if (!vendor) {
+      //   console.log("EventVendor not found!");
+      // } else {
+      //   console.log("Found EventVendor:", vendor);
+      // }
 
-    await prisma.service.create({
-      data: {
-        image_url: req.image_url,
-        service_name: req.service_name,
-        service_description: req.service_description,
-        categories: req.categories,
-        charges: Number(req.charges),
-        event_vendor: {
-          connect: { id: parseInt(request.auth.user.id) },
+      await prisma.service.create({
+        data: {
+          image_url: req.image_url,
+          service_name: req.service_name,
+          service_description: req.service_description,
+          categories: req.categories,
+          charges: Number(req.charges),
+          event_vendor: {
+            connect: { id: parseInt(request.auth.user.id) },
+          },
         },
-      },
-    });
+      });
 
-    return NextResponse.json(
-      { message: "New service created" },
-      { status: 201 }
-    );
+      return NextResponse.json(
+        { message: "New service created" },
+        { status: 201 }
+      );
+    }
   } catch (error) {
     console.error(
       "Error in creating service:",
@@ -47,30 +49,32 @@ export const POST = auth(async function POST(request) {
 
 export const GET = auth(async function GET(request) {
   try {
-    const id = parseInt(request?.auth?.user?.id);
-    // console.log(id)
-    // console.log("dashboard route: ",parseInt( request.auth.user.id));
-    // const vendor = await prisma.eventVendor.findUnique({
-    //   where: { id: parseInt( request.auth.user.id) }
-    // });
-    // if (!vendor) {
-    //   console.log("EventVendor not found!");
-    // } else {
-    //   console.log("Found EventVendor:", vendor);
-    // }
+    if (request.auth.user && request.auth.user.user_type == "vendor") {
+      const id = parseInt(request?.auth?.user?.id);
+      // console.log(id)
+      // console.log("dashboard route: ",parseInt( request.auth.user.id));
+      // const vendor = await prisma.eventVendor.findUnique({
+      //   where: { id: parseInt( request.auth.user.id) }
+      // });
+      // if (!vendor) {
+      //   console.log("EventVendor not found!");
+      // } else {
+      //   console.log("Found EventVendor:", vendor);
+      // }
 
-    const services = await prisma.service.findMany({
-      where: { event_vendor_id: id },
-      orderBy: {
-        service_id: "asc",
-      },
-    });
-    // console.log("vendor services = ", services);
-    return NextResponse.json(
-      { data: services },
-      { message: "Services retrived" },
-      { status: 200 }
-    );
+      const services = await prisma.service.findMany({
+        where: { event_vendor_id: id },
+        orderBy: {
+          service_id: "asc",
+        },
+      });
+      // console.log("vendor services = ", services);
+      return NextResponse.json(
+        { data: services },
+        { message: "Services retrived" },
+        { status: 200 }
+      );
+    }
   } catch (error) {
     console.error(
       "Error in retrieving services:",
@@ -85,26 +89,28 @@ export const GET = auth(async function GET(request) {
 
 export const PUT = auth(async function PUT(request) {
   try {
-    const req = await request.json();
-    console.log(req.categories);
-    if (request.auth.user) {
-      await prisma.service.update({
-        where: {
-          service_id: req.service_id,
-        },
-        data: {
-          ...(req.image_url ? { image_url: req.image_url } : {}),
-          ...(req.service_name ? { service_name: req.service_name } : {}),
-          ...(req.service_description
-            ? { service_description: req.service_description }
-            : {}),
-          ...(req.categories ? { categories: { set: req.categories } } : {}),
-          ...(req.charges ? { charges: parseFloat(req.charges) } : {}),
-        },
-      });
-    }
+    if (request.auth.user && request.auth.user.user_type == "vendor") {
+      const req = await request.json();
+      console.log(req.categories);
+      if (request.auth.user) {
+        await prisma.service.update({
+          where: {
+            service_id: req.service_id,
+          },
+          data: {
+            ...(req.image_url ? { image_url: req.image_url } : {}),
+            ...(req.service_name ? { service_name: req.service_name } : {}),
+            ...(req.service_description
+              ? { service_description: req.service_description }
+              : {}),
+            ...(req.categories ? { categories: { set: req.categories } } : {}),
+            ...(req.charges ? { charges: parseFloat(req.charges) } : {}),
+          },
+        });
+      }
 
-    return NextResponse.json({ message: "Service updated" }, { status: 200 });
+      return NextResponse.json({ message: "Service updated" }, { status: 200 });
+    }
   } catch (error) {
     console.error(
       "Error in delete services:",
@@ -119,17 +125,19 @@ export const PUT = auth(async function PUT(request) {
 
 export const DELETE = auth(async function DELETE(request) {
   try {
-    const req = await request.json();
-    // console.log(request)
-    if (request.auth.user) {
-      await prisma.service.delete({
-        where: {
-          service_id: parseInt(req.service_id),
-        },
-      });
-    }
+    if (request.auth.user && request.auth.user.user_type == "vendor") {
+      const req = await request.json();
+      // console.log(request)
+      if (request.auth.user) {
+        await prisma.service.delete({
+          where: {
+            service_id: parseInt(req.service_id),
+          },
+        });
+      }
 
-    return NextResponse.json({ message: "Service updated" }, { status: 200 });
+      return NextResponse.json({ message: "Service updated" }, { status: 200 });
+    }
   } catch (error) {
     console.error(
       "Error in delete services:",
